Updates for api v 2.1.3

+++ src/main/java/com/exavault/client/model/ResourceDelete.java
@@ -28,7 +28,7 @@ import java.io.IOException;
 
 public class ResourceDelete {
   @SerializedName("id")
-  private Integer id = null;
+  private Long id = null;
 
   /**
    * Type of item. \&quot;resource\&quot;
@@ -76,7 +76,7 @@ public class ResourceDelete {
   @SerializedName("meta")
   private Object meta = null;
 
-  public ResourceDelete id(Integer id) {
+  public ResourceDelete id(Long id) {
     this.id = id;
     return this;
   }
@@ -86,11 +86,11 @@ public class ResourceDelete {
    * @return id
   **/
   @Schema(example = "1", description = "")
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
